models/finetuning/requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
clip @ git+https://github.com/openai/CLIP.git@3b473b0e682c091a9e53623eebc1ca1657385717
configparser==5.0.2
docker-pycreds==0.4.0
ftfy==6.0.3
gitdb==4.0.7
GitPython==3.1.18
idna==3.2
importlib-metadata==4.6.3
joblib==1.0.1
logging==0.4.9.6
nltk==3.6.2
numpy==1.21.1
pathtools==0.1.2
Pillow==8.3.1
promise==2.3
protobuf==3.17.3
psutil==5.8.0
python-dateutil==2.8.2
PyYAML==5.4.1
regex==2021.8.3
requests==2.26.0
scikit-learn==0.24.2
scipy==1.7.1
sentry-sdk==1.3.1
shortuuid==1.0.1
six==1.16.0
sklearn==0.0
smmap==4.0.0
subprocess32==3.5.4
threadpoolctl==2.2.0
torch==1.9.0
torchvision==0.10.0
tqdm==4.62.0
typing-extensions==3.10.0.0
urllib3==1.26.6
wandb==0.10.8
watchdog==2.1.5
wcwidth==0.2.5
zipp==3.5.0
